move themeprovider in App
