+++ src/App.tsx
@@ -1,18 +1,32 @@
 import { Grid } from '@mui/material';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 import Nav from './layout/Nav';
 import SideBar from './layout/SideBar';
 
 function App () {
-  return (
-    <Grid container spacing={1}>
-      <Grid item xs={12}>
-        <Nav />
-      </Grid>
+  const theme = createTheme({
+    palette: {
+      primary: {
+        main: '#E1B15D', // brown
+      },
+      secondary: {
+        main: '#FFFFFF', // white
+      },
+    },
+  });
 
-      <Grid item xs={2}>
-        <SideBar />
+  return (
+    <ThemeProvider theme={theme}>
+      <Grid container spacing={1}>
+        <Grid item xs={12}>
+          <Nav />
+        </Grid>
+
+        <Grid item xs={2}>
+          <SideBar />
+        </Grid>
       </Grid>
-    </Grid>
+    </ThemeProvider>
   );
 }
